Implement DRY principle for rendering blocks
