+++ src/lib/components/BodyRenderer/Blocks/tableCell.svelte
@@ -1,19 +1,8 @@
 <script lang="ts">
 	/* eslint-disable @typescript-eslint/no-unused-vars */
-	import { Switch, Case } from '$components/Switch';
-
-	import InlineBlock from '../inline.svelte';
-	import TextWrapper from './textWrapper.svelte';
-	import NestedList from './nestedList.svelte';
-	import Image from './image.svelte';
-	import Embed from './embed.svelte';
-	import TallyButton from './tallyButton.svelte';
-	import Chart from './chart.svelte';
-	import TwitterEmbed from './twitterEmbed.svelte';
+	import Item from '../item.svelte'
 
 	export let cell: any;
 </script>
 
-<template lang="pug" src="./tableCell.pug">
-
-</template>
+<Item item={cell} />
